+++ samples/v1.22.2/React/src/webparts/helloWorld/loc/qps-ploc.js
@@ -1,0 +1,16 @@
+define([], function() {
+  return {
+    "PropertyPaneDescription": "[!!Ďēśċŕîƥţîōń!!]",
+    "BasicGroupName": "[!!Ĝŕōûƥ Ńàmē!!]",
+    "DescriptionFieldLabel": "[!!!Ďēśċŕîƥţîōń Ƒîēļď!!!]",
+    "AppLocalEnvironmentSharePoint": "[!!!!!!!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ ōń ŷōûŕ ļōċàļ ēńvîŕōńmēńţ àś ŚĥàŕēƤōîńţ wēƀ ƥàŕţ!!!!!!!!!!!!]",
+    "AppLocalEnvironmentTeams": "[!!!!!!!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ ōń ŷōûŕ ļōċàļ ēńvîŕōńmēńţ àś Mîċŕōśōƒţ Ţēàmś àƥƥ!!!!!!!!!!!!]",
+    "AppLocalEnvironmentOffice": "[!!!!!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ ōń ŷōûŕ ļōċàļ ēńvîŕōńmēńţ îń ōƒƒîċē.ċōm!!!!!!!!!!!]",
+    "AppLocalEnvironmentOutlook": "[!!!!!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ ōń ŷōûŕ ļōċàļ ēńvîŕōńmēńţ îń Ōûţļōōķ!!!!!!!!!!]",
+    "AppSharePointEnvironment": "[!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ ōń ŚĥàŕēƤōîńţ ƥàĝē!!!!!!!]",
+    "AppTeamsTabEnvironment": "[!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ îń Mîċŕōśōƒţ Ţēàmś!!!!!!!]",
+    "AppOfficeEnvironment": "[!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ îń ōƒƒîċē.ċōm!!!!!!]",
+    "AppOutlookEnvironment": "[!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ îń Ōûţļōōķ!!!!!!]",
+    "UnknownEnvironment": "[!!!!!!!!Ţĥē àƥƥ îś ŕûńńîńĝ îń àń ûńķńōwń ēńvîŕōńmēńţ!!!!!!!!]"
+  }
+});